flask/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.24.3
pandas==1.5.3
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tqdm==4.65.0
Werkzeug==2.3.4
zipp==3.15.0
uWSGI
